Extract a hook and context, and use them for app initialization
