+++ react-native-expo/components/RootLayout/index.tsx
@@ -1,5 +1,13 @@
+import AppInitializationContext from '@/contexts/AppInitializationContext';
 import { Stack } from 'expo-router';
+import useInitializeApp from './useInitializeApp';
 
 export default function RootLayout() {
-  return <Stack />;
+  const appInitialization = useInitializeApp();
+
+  return (
+    <AppInitializationContext.Provider value={appInitialization}>
+      <Stack />
+    </AppInitializationContext.Provider>
+  );
 }
