+++ react-native-expo/contexts/AppInitializationContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+const AppInitializationContext = createContext<{ isLoading: boolean; error: string | null }>({
+  isLoading: false,
+  error: null,
+});
+export default AppInitializationContext;
